fix: adds ability to split a text via multiple line ending characters
different platforms use different line endings based on encoding. `\r`
and `\n` are both valid and should both be checked.

changes the `NEW_LINE` static string to be an array of strings
`NEW_LINES` and creates a new function `splitLines` which will use all
the characters in `NEW_LINES`.

in the case that it wants to rebuild text and use a line break it will
default to the first character `NEW_LINES[0]` which is the `\n` char